Fixed bugs in wss::Path. Fixes unit tests.

+++ wss/gameserver/include/wss/Utils.h
@@ -5,21 +5,21 @@ namespace wss {
 class Utils {
 public:
 
-	static inline size_t XYToIndex(const size_t x, const size_t y, const size_t width) {
+	static inline size_t XYToIndex(const int x, const int y, const size_t width) {
 	return y * width + x;
 	}
 
 	static inline size_t XYToIndex(const glm::vec2 &position, const size_t width) {
-		return (size_t)(position.y) * width + size_t(position.x);
+		return size_t((position.y) * width + size_t(position.x));
 	}
 
 
 	static inline void indexToXY(const size_t index, const size_t width, glm::vec2 &position) {
-		position.x = index % width;
-		position.y = index / width;
+		position.x = int(index % width);
+		position.y = int(index / width);
 	}
 
-	static inline void indexToXY(const size_t index, const size_t width, size_t &x, size_t &y) {
+	static inline void indexToXY(const size_t index, const size_t width, int &x, int &y) {
 		x = index % width;
 		y = index / width;
 	}
